+++ GuitarHarmonics/InfiniteHarmonics.js
@@ -1,0 +1,10 @@
+import { Primitive } from "../sketchlib/primitives/Primitive.js";
+
+/**
+ * @implements {Primitive}
+ */
+export class InfiniteHarmonics {
+  draw(p) {
+    throw new Error("Method not implemented.");
+  }
+}
